[refactor] 불필요한 코드 수정
