+++ README.md
@@ -3,8 +3,8 @@
 
 ## 기능 구현 목록
 ### 체스 1단계
-- [ ] 체스 게임을 할 수 있는 체스판을 초기화 하는 기능
-    - [ ] 가로 위치를 왼쪽부터 a ~ h로 설정
-    - [ ] 세로 위치를 아래부터 1 ~ 8로 설정
-    - [ ] 검은 말은 위에서부터 대문자로 구분하고,흰 말은 아래서부터 소문자로 구분 
+- [x] 체스 게임을 할 수 있는 체스판을 초기화 하는 기능
+    - [x] 가로 위치를 왼쪽부터 a ~ h로 설정
+    - [x] 세로 위치를 아래부터 1 ~ 8로 설정
+    - [x] 검은 말은 위에서부터 대문자로 구분하고,흰 말은 아래서부터 소문자로 구분 
 - [ ] 초기화 된 체스판을 출력하는 기능
