+++ src/main/java/chess/domain/chessBoard/ChessBoard.java
@@ -1,8 +1,8 @@
-package chess;
+package chess.domain.chessBoard;
 
-import chess.domain.chessPiece.*;
+import chess.domain.chessPiece.Piece;
+import chess.domain.position.Position;
 
-import java.util.HashMap;
 import java.util.Map;
 
 public class ChessBoard {
