+++ src/main/java/chess/domain/chessPiece/pieceType/PieceType.java
@@ -1,4 +1,4 @@
-package chess.domain.chessPiece;
+package chess.domain.chessPiece.pieceType;
 
 public interface PieceType {
     String getName();
