+++ src/test/java/chess/domain/chessPiece/PieceColorTest.java
@@ -1,8 +1,8 @@
-package chess;
+package chess.domain.chessPiece;
 
-import chess.domain.chessPiece.King;
+import chess.domain.chessPiece.pieceType.King;
 import chess.domain.chessPiece.PieceColor;
-import chess.domain.chessPiece.PieceType;
+import chess.domain.chessPiece.pieceType.PieceType;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
