+++ src/test/java/chess/domain/position/ChessRankTest.java
@@ -1,5 +1,6 @@
-package chess;
+package chess.domain.position;
 
+import chess.domain.position.ChessRank;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
